// packages/functions/src/hasura-metadata/types.ts
export interface HasuraMetadataAPIPayload<Type extends string, Args> {
  type: Type
  args: Args
}

/**
 * @see{@link https://hasura.io/docs/latest/api-reference/syntax-defs/#headerfromvalue}
 */
export interface HasuraHeaderFromValue {
  /**
   * Name of the header
   */
  name: string
  /**
   * Value of the header
   */
  value: string
}

/**
 * @see{@link https://hasura.io/docs/latest/api-reference/syntax-defs/#headerfromenv}
 */
export interface HasuraHeaderFromEnv {
  /**
   * Name of the header
   */
  name: string
  /**
   * Name of the environment variable which holds the value of the header
   */
  value_from_env: string
}

export type HasuraScheduledEventHeaders = (HasuraHeaderFromValue | HasuraHeaderFromEnv)[]

/**
 * Payload used to create a scheduled event
 * @see{@link https://hasura.io/docs/latest/api-reference/metadata-api/scheduled-triggers/#metadata-create-scheduled-event}
 */
export interface HasuraCreateScheduledEventPayload<T>
  extends HasuraMetadataAPIPayload<
    'create_scheduled_event',
    {
      /**
       * A String value which supports templating environment variables enclosed in {{ and }}.
       * @example
       * https://{{ACTION_API_DOMAIN}}/create-user
       */
      webhook: string
      /**
       * The time at which the invocation should be invoked. (ISO8601 format)
       */
      schedule_at: string
      /**
       * Any JSON payload which will be sent when the webhook is invoked.
       */
      payload?: T & Record<string, unknown> // TODO check tyoe

      /**
       * List of headers to be sent with the webhook
       */
      headers?: (HasuraHeaderFromValue | HasuraHeaderFromEnv)[]
      /**
       * Retry configuration if scheduled event delivery fails
       */
      retry_conf?: {
        /**
         * Number of times to retry delivery.
         * @default 0
         */
        num_retries?: number
        /**
         * Number of seconds to wait between each retry.
         * @default 10
         */
        timeout_seconds?: number
        /**
         * Number of seconds to wait for response before timing out
         * @default 60
         */
        tolerance_seconds?: number
        /**
         * Number of seconds between scheduled time and actual delivery time that is acceptable.
         * If the time difference is more than this, then the event is dropped.
         * @default 21600 (6 hours)
         */
        retry_interval_seconds?: number
      }
      /**
       * Custom comment.
       */
      comment?: string
    }
  > {}

export interface HasuraCreateScheduledEventResult {
  message: 'success'
  /**
   * UUID of the scheduled event
   */
  event_id: string
}

/**
 * @see{@link https://hasura.io/docs/latest/api-reference/metadata-api/scheduled-triggers/#metadata-delete-scheduled-event}
 */
export interface HasuraDeleteScheduledEventPayload
  extends HasuraMetadataAPIPayload<
    'delete_scheduled_event',
    {
      /**
       * Type of the event trigger.
       */
      type: 'one_off' | 'cron'
      /**
       * The id of the scheduled event.
       */
      event_id: string
    }
  > {}

export type HasuraMetadataPayload<T = any> =
  | HasuraCreateScheduledEventPayload<T>
  | HasuraDeleteScheduledEventPayload